AddedBlur

+++ Assets/Scripts/Ability/ArcherAbilities/Blur/Blur.cs
@@ -3,5 +3,5 @@ using UnityEngine;
 [CreateAssetMenu(fileName = "Blur", menuName = "Ability/Range/Blur")]
 public class Blur : Ability
 {
-    [field: SerializeField] public int Evasion {  get; private set; }   
+    [field: SerializeField] public float Evasion {  get; private set; }   
 }
